+++ packages/jest-runtime/src/__mocks__/createRuntime.js
@@ -8,10 +8,10 @@
 import path from 'path';
 
 module.exports = async function createRuntime(filename, config) {
-  const NodeEnvironment = require('jest-environment-node');
-  const Runtime = require('../');
+  const {default: NodeEnvironment} = await import('jest-environment-node');
+  const {default: Runtime} = await import('../');
 
-  const {normalize} = require('jest-config');
+  const {normalize} = await import('jest-config');
 
   config = normalize(
     {