+++ CHANGELOG.md
@@ -4,6 +4,7 @@
 
 - `[jest-circus, jest-config, jest-runtime]` Add new `injectGlobals` config and CLI option to disable injecting global variables into the runtime ([#10484](https://github.com/facebook/jest/pull/10484))
 - `[jest-each]` Fixes `.each` type to always be callable ([#10447](https://github.com/facebook/jest/pull/10447))
+- `[jest-runner]` Add support for `moduleLoader`s with `default` exports ([#10541](https://github.com/facebook/jest/pull/10541))
 - `[jest-haste-map]` Add support for .mock extension for manual mocks ([#10460](https://github.com/facebook/jest/pull/10460))
 
 ### Fixes